Split up index.js into api and snapshot
